third_party/WebKit/LayoutTests/editing/selection/modify_move/move-by-character-004.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<div contenteditable id="root" class="editing">
<span id="sample">a<img src="../../resources/abe.png">new 
<br>nation <i>  </i> <img src="../../resources/abe.png"> conceived
<br>nation<img src="../../resources/abe.png"></span>
</div>
<div id="log"></div>
<script>
test(function () {
    var selection = window.getSelection();
    selection.collapse(sample, 0);
    for (i = 0; i < 33; ++i)
    selection.modify('move', 'forward', 'character');
    for (i = 0; i < 32; ++i)
    selection.modify('move', 'backward', 'character');
    assert_true(selection.isCollapsed, 'isCollapsed');
    assert_equals(selection.anchorNode, sample.firstChild, 'anchorNode');
    assert_equals(selection.anchorOffset, 1, 'anchorOffset');
});
</script>

</body>
